move contents of context_menu_decoy_child to context_menu to fix Linux analysis
