+++ dev/docs/footer.html
@@ -1,0 +1,1 @@
+<script src="footer.js"></script>
