+++ packages/flutter/test/cupertino/context_menu_test.dart
@@ -7,7 +7,6 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
-import 'package:flutter/src/cupertino/context_menu_decoy_child.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
